refactor isAppReady out of state/store
