+++ dash-renderer/src/reducers/reducer.js
@@ -10,7 +10,6 @@ import {
     view,
 } from 'ramda';
 import {combineReducers} from 'redux';
-import isAppReady from './isAppReady';
 import layout from './layout';
 import graphs from './dependencyGraph';
 import paths from './paths';
@@ -32,7 +31,6 @@ export const apiRequests = [
 function mainReducer() {
     const parts = {
         appLifecycle,
-        isAppReady,
         layout,
         graphs,
         paths,
